@@ -1,12 +1,11 @@
 name: redux_api_middleware
 description: A dart redux middleware for calling APIs. APIs interact with reducers made easy..
 homepage: https://github.com/carloslira/redux_api_middleware
-version: 0.2.1
+version: 0.3.0
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  meta: ">=1.0.0 <2.0.0"
-  redux: ">=4.0.0 <5.0.0"
-  http: ^0.12.0+4
+  redux: ">=5.0.0 <6.0.0"
+  http: ">=0.13.0 <0.14.0"
